fix: update imports for Certification entity in Freelancer and User specs refactor: adjust constructor parameters in CertificationService
