+++ src/contexts/CoreContext/domain/services/CertificationService.ts
@@ -1,4 +1,3 @@
-import { ICertificationService } from "../interfaces/services/ICertificationService";
 import { Certification } from "../entities/Certification";
 
 export class CertificationService {
