+++ tests/contexts/coreContext/Freelancer.spec.ts
@@ -7,7 +7,7 @@ import { ExperienceService } from "@/contexts/CoreContext/domain/services/Experi
 import { LanguageService } from "@/contexts/CoreContext/domain/services/LanguageService";
 import { SkillService } from "@/contexts/CoreContext/domain/services/SkillService";
 import { About } from "@/contexts/CoreContext/domain/valueObjects/About";
-import { Certification } from "@/contexts/CoreContext/domain/valueObjects/Certification";
+import { Certification } from "@/contexts/CoreContext/domain/entities/Certification";
 import { Language } from "@/contexts/CoreContext/domain/valueObjects/Language";
 import { Skill } from "@/contexts/CoreContext/domain/valueObjects/Skill";
 import { UserId } from "@/contexts/CoreContext/domain/valueObjects/UserId";
